refactor(generation): move poseidon2 generation code
This is 100% copy and paste poseidon2 generation code into its specific
directories, as part of an effort to delete `generation` and move trace
gen code into their specific kind of table.
