+++ circuits/src/generation/mod.rs
@@ -11,9 +11,6 @@ pub mod io_memory;
 pub mod memory;
 pub mod memory_zeroinit;
 pub mod memoryinit;
-pub mod poseidon2;
-pub mod poseidon2_output_bytes;
-pub mod poseidon2_sponge;
 pub mod program;
 pub mod rangecheck;
 pub mod rangecheck_u8;
@@ -44,8 +41,6 @@ use self::memoryinit::{
     generate_call_tape_init_trace, generate_event_tape_init_trace, generate_memory_init_trace,
     generate_private_tape_init_trace, generate_public_tape_init_trace,
 };
-use self::poseidon2_output_bytes::generate_poseidon2_output_bytes_trace;
-use self::poseidon2_sponge::generate_poseidon2_sponge_trace;
 use self::rangecheck::generate_rangecheck_trace;
 use self::rangecheck_u8::generate_rangecheck_u8_trace;
 use self::xor::generate_xor_trace;
@@ -57,8 +52,10 @@ use crate::generation::memory_zeroinit::generate_memory_zero_init_trace;
 use crate::generation::memoryinit::{
     generate_elf_memory_init_trace, generate_mozak_memory_init_trace,
 };
-use crate::generation::poseidon2::generate_poseidon2_trace;
 use crate::generation::program::generate_program_rom_trace;
+use crate::poseidon2::generation::generate_poseidon2_trace;
+use crate::poseidon2_output_bytes::generation::generate_poseidon2_output_bytes_trace;
+use crate::poseidon2_sponge::generation::generate_poseidon2_sponge_trace;
 use crate::register::generation::{generate_register_init_trace, generate_register_trace};
 use crate::stark::mozak_stark::{
     all_starks, MozakStark, PublicInputs, TableKindArray, TableKindSetBuilder,
