+++ circuits/src/poseidon2/mod.rs
@@ -1,2 +1,3 @@
 pub mod columns;
+pub mod generation;
 pub mod stark;
