+++ circuits/src/poseidon2_output_bytes/mod.rs
@@ -1,2 +1,3 @@
 pub mod columns;
+pub mod generation;
 pub mod stark;
